fix colors in app
