+++ lib/src/app/view/task_tracking_app.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:task_tracking_app/src/app/view/main_view.dart';
-import 'package:task_tracking_app/src/app/view/todo_list.dart';
 
 
 class TaskTrackingApp extends StatelessWidget {
@@ -11,8 +10,7 @@ class TaskTrackingApp extends StatelessWidget {
     return new MaterialApp(
           title: appName,
           home: new MainView(title: appName,)
-//        title: 'Daily task list',
-//        home: new TodoList()
+
     );
   }
 }
